refactoring: moved graph representation, deleted playground file
